Added tapir for desk.com integration.
Wrote a small library for quickly defining and reading from an API.  It's modeled after the desk.com one I found at first, but is lighter so it should be quicker to update with the api calls I need.  Also it uses curl for transport instead of buzz, so I won't run into the file_get_ocntents problem.