Fixed desk api's case->update.  #1
Figured out the problem.  custom_fields must NOT be in the query.  I knew they had to be part of the post data, but didn't realize that if they were in the query it would be invalid.

Code still needs a cleanup.  I wanted to get a working state into git.