Added desk feedback storage.  openscholar/openscholar#3247
Added a menu entry for depositing user feedback.  When a desk issue is closed, its automatic email should include links for rating the support as Excellent, Good, Fair, or Poor.  Those links point to my new menu entry.

The link also includes the timestamp that the entry was last updated (which is going to be final, since the ticket is clsoed at this point) and the id of the case.  Integrated support fetches the case itself from desk by id and compares the timestamps.  This should be relatively hard for someone to spoof.

Ratings are saved immediately.  Users can also attach comments.  I saved the ratings right away just in case someone gets put off by the form.