Getsatisfaction tickets sync to github openscholar/openscholar#3169
Wrote a cron job that gets all GS tickets with a status of pending.  If they aren't already linked to a github ticket, create a new ticket and link them together.

Haven't tested this in a cron context, but it works outside of that.

One major todo.  I haven't figured out how to handle paged requests.  GS will give me the first 50 or 100 pending issues.  Integrated support should be smart enough to loop to the next set until it's gone over all of them.