Fixed getsatisfaction topic pushes openscholar/openscholar#3169
Added setting for whether or not to enforce same authorship.  This'll let us work with it for now and easily change things when GS tells me wha'ts wrong with their api.

Stepped through entire pending GS topic -> github process locally.  Now to do it in dev.

Labels work like this.  Anything that's mapped becomes pending.  Anything with status: in dev, needs qa, or needs review becomes active.  And anything that's closed is marked as complete.

There are two other possible GS statuses.  I can use null to strip a status, but I don't know why I'd want to do that programatically.  THere's also rejected, but that's the opposite of sending a ticket to GH for development.