Github applies desk label openscholar/openscholar#3168
When desk created github tickets, it gave them a "desk" label.  But if you added a github id to a desk ticket, no label was added.  Now that label's existence is checked, and the label is added if it's missing.

I haven't tested the hook (since localhost isn't receiving github ewbhooks) but the add label api call works perfectly.