GS <-> GH link creates GH comment openscholar/openscholar#3169
When a link is created to map a GS topic ot a GH issue, the GH issue will receive a comment with an href to the GS topic.  I removed the old code for manually triggering an update - GH receiving the comment will send a webhook and do that part for us.