Roadmap labels link admins to gh openscholar/openscholar#3355
The roadmap should be one stop shopping for people trying to see what the current issues in OS are.  But it shouldn't encourage anyone to dump unwanted comments into issues that are being worked on.  The solution is that admins get a GH link.  It's in teh label.  (Yes this means that tickets that aren't started yet don't have a working link.  Until development starts, we should be looking at the GH issue anyway)