Try/catch around github exceptions openscholar/openscholar#3355
Github api was getting 500 server errors.  Wrapped them in an exception.

For issues on the roadmap, the exceptions are being logged, but the rest of the roadmap will display.

For milestones, we can't really display the roadmap without milestones.  Instead there's a drupal error.  Because this is a bit more severe, I'm making a 1 hour cache of milestones whne they're fetched correctly.