Began OSTrello.  It GETs and POSTs.
OSTrello is a wrapper for the Trello class.  It has a full model of the API endpoints for Trello.  It allows you to call the different endpoints as functions of the client object, passing ids and parameters at each call.

For example:

$trello->boards('abc123')->lists()->get(); // gets all the lists on the board identified by abc123
$trello->cards()->create(array('idList'=>'myListId', 'name' => 'this is the title of my card')); // creates a new card on tehe spcified list.

Get and create are working.  Update and delete need to be written, but that should be easy since they're just wrapping the original lib.