Cleaned up trello model.
A url endpoint can be a string in an array.  It does not have to have its own array.  If it does, the key will be used as the url, and the array will define children.  This saved a ton of code and cost only 4 lines to check what was what.