Began processing trello events #96
Started logging and looking at trello events.  Caught the event for list movement.  When on the roadmap board, if a card is moved to a list whose name matches a milestone, github's milestone will be updated.

Also caught event for moving between boards.  Does not include the names of the lists, so I can't apply an engineering/bug/idea label at this point.  Instead went back and added that label at card creation time.  Still have to figure out how to tell if a GH ticket needs to be created at this point or not.