Trello can push labels.  #96
The labels thing was bothering me.  The api doesn't let me put the Features label on a card.  I can only put a color on a card and hope it's the right one.

I made a var mapping labels to colors.  Added a checkbox to the admin page.  Checking it will set the trello labels on the roadmap page to match the colors in the var.