Trello roadmap uses ajax #96
The trello roadmap now uses ajax to fetch statuses from github.  I'm still using my big queries instead of looping over tickets, but this should speed up page loads even more.

I also cached the statuses.  When the ajax call builds its array of status => [list,of,ids], it caches the whole thing for 30 minutes.  When the roadmap table is built, if the cached data is valid it gets provided to Drupal.settings.  When that setting is present and nonempty, javascript will use it instead of ajax.