Prevents GS from linking dupes openscholar/openscholar#3657
Changed the validation function.  It used to check topic/service/id to avoid duplicate rows.  I removed id, so now it's just checking that the topic isn't already mapped on the service.  All additions to this table go through the form, so validation should work when links are added programmatically.