Backlog tab has different cols #96
Backlog lost hotfix and gained projects.

Found a bug in backlog while I was in there.  It wasn't getting all the items - I was still operating on a single list instead of the whole board.  Fixed it.

Also removed the current milestone label since it was redundant with select list.